Updates for days.
BattleField refactored to Battlefield.
PlayerTanks now stored in a Map with the players' UUID as the key.
Battlefields now stored in a Map with their name as the key.
Moved PlayerTank to a separate package from the battlefield classes.
Menus only open if the player is in a battlefield.
Implemented Money and XP statistics.
Removed some extra loops that were happening.
Check DamageCause in onEntityDamage method.
Updated TODO list.
Implemented custom health system for balancing purposes.
Added weight to tanks to help track melee damage.